New trainer scoring
Should result in better scoring performance when game's nature is added to play, assuming I didn't introduce bugs.

+++ mltrainer/src/adapter/mod.rs
@@ -10,5 +10,5 @@ where
 }
 
 pub trait TrainerAdapter: Sync {
-    fn try_out(&self, network: &mut Network) -> f32;
+    fn try_out(&self, network: &mut Network) -> isize;
 }
